consolidate clang-tidy suppressions
- enable several disabled clang-tidy checks, and fix associated warnings
	- cppcoreguidelines-avoid-magic-numbers
	- hicpp-signed-bitwise
	- readability-else-after-return
	- readability-magic-numbers
- centralize suppressions for bugprone-reserved-identifier and thereby
  avoid NOLINTs for __asan_default_options(), __llvm_profile_*, etc
- replace NOLINTs for *-narrowing-conversions with casts
- replace NOLINTs for va_list declarations with zero-init {0}
- avoid NOLINTs for forward-declared proto structs
- avoid NOLINTs for clang-analyzer-deadcode.DeadStores, either by using
  the variable(s) in question explicitly or by refactoring such that
  clang-analyzer understands the variable usage in question
- remove mutable global variable in test along with associated NOLINT
- simplify NOLINTBEGIN/NOLINTEND into NOLINT where possible

+++ src/sandbox/linux/landlock.c
@@ -1,7 +1,7 @@
 #ifndef _GNU_SOURCE
-#define _GNU_SOURCE // NOLINT(bugprone-reserved-identifier)
+#define _GNU_SOURCE
 #endif
-#define __SANE_USERSPACE_TYPES__ // NOLINT(bugprone-reserved-identifier)
+#define __SANE_USERSPACE_TYPES__
 
 /*
  * Some helpful Landlock references:
@@ -23,11 +23,7 @@ landlock_create_ruleset(const struct landlock_ruleset_attr *const attr,
                         const size_t size,
                         const __u32 flags)
 {
-	return syscall( // NOLINT(*-narrowing-conversions)
-		__NR_landlock_create_ruleset,
-		attr,
-		size,
-		flags);
+	return (int)syscall(__NR_landlock_create_ruleset, attr, size, flags);
 }
 #endif
 
@@ -38,12 +34,11 @@ landlock_add_rule(const int ruleset_fd,
                   const void *const rule_attr,
                   const __u32 flags)
 {
-	return syscall( // NOLINT(*-narrowing-conversions)
-		__NR_landlock_add_rule,
-		ruleset_fd,
-		rule_type,
-		rule_attr,
-		flags);
+	return (int)syscall(__NR_landlock_add_rule,
+	                    ruleset_fd,
+	                    rule_type,
+	                    rule_attr,
+	                    flags);
 }
 #endif
 
@@ -51,10 +46,7 @@ landlock_add_rule(const int ruleset_fd,
 static inline int
 landlock_restrict_self(const int ruleset_fd, const __u32 flags)
 {
-	return syscall( // NOLINT(*-narrowing-conversions)
-		__NR_landlock_restrict_self,
-		ruleset_fd,
-		flags);
+	return (int)syscall(__NR_landlock_restrict_self, ruleset_fd, flags);
 }
 #endif
